Change the behaviour of scale
Scale is typically used to scale up small videos (old cameras) or small regions of interest. Often, the goal is to make the image "big enough" (the actual scale up does not matter as long as the image is large enough to avoid issues with labels and playback). Whilst the new implementation now does this by default, the old behaviour (specify the actual scale factor) is still supported.