Aedat compatibility with DV
Multiple changes to ensure that generated aedat files can be loaded by DV. The biggest change is the addition of file data tables (essentially a data index at the end of aedat files), which DV requires.

Whilst loading and play back works, I've experienced minor issues when trying to load multiple files in a row (restarting DV is often required), when playing short files, or when enabling the loop playback function of DV. Some of these issues could come from the size of the packets generated by faery (typically bigger than DV's), a bug in faery.aedat, or a bug in DV. Since the latter is not completely open-source, we will need help from Inivation / Synsense to investigate further (if there is interest for better support).

Fixes #13.